+++ lib/lotus/migrations/postgres/v1.ex
@@ -21,6 +21,7 @@ defmodule Lotus.Migrations.Postgres.V1 do
       add(:description, :text)
       add(:query, :map, null: false)
       add(:tags, {:array, :string}, default: [])
+      add(:data_repo, :string)
 
       timestamps(type: :utc_datetime_usec)
     end