+++ lib/lotus/migrations/sqlite.ex
@@ -13,6 +13,7 @@ defmodule Lotus.Migrations.SQLite do
       add(:description, :text)
       add(:query, :map, null: false)
       add(:tags, :text, default: "[]")
+      add(:data_repo, :string)
       timestamps(type: :utc_datetime_usec)
     end